+++ apps/rig_api/lib/rig_api/controllers/apis_controller.ex
@@ -242,7 +242,26 @@ defmodule RigApi.ApisController do
 
           properties do
             id(:string, "Endpoint ID", required: true, example: "get-auth-register")
-            path(:string, "Endpoint path", required: true, example: "/auth/register")
+
+            path(:string, "Endpoint path. Curly braces may be used to ignore parts of the path.",
+              required: false,
+              example: "/auth/register/{user}"
+            )
+
+            path_regex(
+              :string,
+              "Endpoint path, given as a regular expression (note that JSON requires escaping backslash characters).",
+              required: false,
+              example: "/auth/register/(.+)"
+            )
+
+            path_replacement(
+              :string,
+              "If given, the request path is rewritten. When used with `path_regex`, capture groups can be referenced by number (note that JSON requires escaping backslash characters).",
+              required: false,
+              example: ~S"/auth/register/\1"
+            )
+
             method(:string, "Endpoint HTTP method", required: true, example: "GET")
             not_secured(:boolean, "Endpoint Security", default: true, example: true)
           end