+++ apps/rig_inbound_gateway/lib/rig_inbound_gateway/api_proxy/handler/kafka.ex
@@ -53,10 +53,10 @@ defmodule RigInboundGateway.ApiProxy.Handler.Kafka do
   # ---
 
   @impl Handler
-  def handle_http_request(conn, api, endpoint)
+  def handle_http_request(conn, api, endpoint, request_path)
 
   @doc "CORS response for preflight request."
-  def handle_http_request(%{method: "OPTIONS"} = conn, _, %{"target" => "kafka"}) do
+  def handle_http_request(%{method: "OPTIONS"} = conn, _, %{"target" => "kafka"}, _) do
     conn
     |> with_cors()
     |> Conn.send_resp(:no_content, "")
@@ -66,7 +66,8 @@ defmodule RigInboundGateway.ApiProxy.Handler.Kafka do
   def handle_http_request(
         %{params: %{"partition" => partition, "event" => event}} = conn,
         _,
-        %{"target" => "kafka"} = endpoint
+        %{"target" => "kafka"} = endpoint,
+        request_path
       ) do
     kafka_message =
       event
@@ -78,7 +79,7 @@ defmodule RigInboundGateway.ApiProxy.Handler.Kafka do
         scheme: conn.scheme,
         headers: Enum.map(conn.req_headers, &Tuple.to_list(&1)),
         method: conn.method,
-        path: conn.request_path,
+        path: request_path,
         query: conn.query_string
       })
       |> Poison.encode!()
@@ -98,7 +99,7 @@ defmodule RigInboundGateway.ApiProxy.Handler.Kafka do
     end
   end
 
-  def handle_http_request(conn, _, %{"target" => "kafka"}) do
+  def handle_http_request(conn, _, %{"target" => "kafka"}, _) do
     response = """
     Bad request: missing expected body parameters.